app-monitoring/requirements.txt
kafka-python
docker
paramiko==2.4.2
